requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
httpx>=0.25.0
validators>=0.22.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pydantic>=2.4.0 